+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - [+] add `and()`, `or()` and `not()` functions
 - [+] add `all()` and `any()` functions
 - [+] add `test()` and `replace()` functions
+- [+] add `slice()` and `slice2()` functions
 
 
 ## 0.1.0 (10.01.2018)